Refactor config manager and added new gem
