+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>ne.fnfal113</groupId>
     <artifactId>FNAmplifications</artifactId>
-    <version>3.2.0-UNOFFICIAL</version>
+    <version>3.2.1-UNOFFICIAL</version>
     <packaging>jar</packaging>
 
     <name>FNAmplifications</name>
