+++ src/main/java/ne/fnfal113/fnamplifications/gears/abstracts/AbstractGears.java
@@ -4,16 +4,35 @@ import io.github.thebusybiscuit.slimefun4.api.items.ItemGroup;
 import io.github.thebusybiscuit.slimefun4.api.items.SlimefunItem;
 import io.github.thebusybiscuit.slimefun4.api.items.SlimefunItemStack;
 import io.github.thebusybiscuit.slimefun4.api.recipes.RecipeType;
+import ne.fnfal113.fnamplifications.FNAmplifications;
 import org.bukkit.entity.Player;
 import org.bukkit.event.entity.EntityDamageByEntityEvent;
 import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.meta.ItemMeta;
 
+import java.io.IOException;
 import java.util.List;
 
 public abstract class AbstractGears extends SlimefunItem {
 
     public AbstractGears(ItemGroup itemGroup, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
         super(itemGroup, item, recipeType, recipe);
+
+        try {
+            FNAmplifications.getInstance().getConfigManager()
+                    .setBooleanValues(this.getId(), "unbreakable", false, "fn-gear-unbreakable-settings");
+            setUnbreakable();
+        } catch (IOException e){
+            FNAmplifications.getInstance().getLogger()
+                    .info("An error has occurred upon setting the armor unbreakable settings! Please report on github!");
+            e.printStackTrace();
+        }
+    }
+
+    public final void setUnbreakable() {
+        ItemMeta meta = this.getItem().getItemMeta();
+        meta.setUnbreakable(FNAmplifications.getInstance().getConfigManager().getBoolById(this.getId(), "unbreakable"));
+        this.getItem().setItemMeta(meta);
     }
 
     /**
