+++ src/main/java/ne/fnfal113/fnamplifications/gems/LifestealGem.java
@@ -4,23 +4,27 @@ import io.github.thebusybiscuit.slimefun4.api.items.ItemGroup;
 import io.github.thebusybiscuit.slimefun4.api.items.SlimefunItem;
 import io.github.thebusybiscuit.slimefun4.api.items.SlimefunItemStack;
 import io.github.thebusybiscuit.slimefun4.api.recipes.RecipeType;
-import ne.fnfal113.fnamplifications.gems.handlers.GemUpgrade;
-import ne.fnfal113.fnamplifications.gems.implementation.Gem;
+import ne.fnfal113.fnamplifications.FNAmplifications;
 import ne.fnfal113.fnamplifications.gems.abstracts.AbstractGem;
-import ne.fnfal113.fnamplifications.utils.WeaponArmorEnum;
+import ne.fnfal113.fnamplifications.gems.handlers.GemUpgrade;
 import ne.fnfal113.fnamplifications.gems.handlers.OnDamageHandler;
+import ne.fnfal113.fnamplifications.gems.implementation.Gem;
 import ne.fnfal113.fnamplifications.utils.Utils;
+import ne.fnfal113.fnamplifications.utils.WeaponArmorEnum;
+import org.bukkit.Bukkit;
+import org.bukkit.attribute.Attribute;
 import org.bukkit.entity.LivingEntity;
 import org.bukkit.entity.Player;
 import org.bukkit.event.entity.EntityDamageByEntityEvent;
 import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.inventory.ItemStack;
 
+import java.util.Objects;
 import java.util.concurrent.ThreadLocalRandom;
 
-public class ThunderBoltGem extends AbstractGem implements OnDamageHandler, GemUpgrade {
+public class LifestealGem extends AbstractGem implements OnDamageHandler, GemUpgrade {
 
-    public ThunderBoltGem(ItemGroup itemGroup, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
+    public LifestealGem(ItemGroup itemGroup, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
         super(itemGroup, item, recipeType, recipe, 12);
     }
 
@@ -55,13 +59,20 @@ public class ThunderBoltGem extends AbstractGem implements OnDamageHandler, GemU
         }
 
         Player player = (Player) event.getDamager();
+
         LivingEntity livingEntity = (LivingEntity) event.getEntity();
 
         if(ThreadLocalRandom.current().nextInt(100) < getChance() / getTier(itemStack, this.getId())){
-            livingEntity.getWorld().strikeLightning(livingEntity.getLocation());
-            player.setNoDamageTicks(20);
-            sendGemMessage(player, this.getItemName());
-        } // if below the chance, strike lightning at the victim and set no damage for the attacker for 1 second
+            Bukkit.getScheduler().runTaskLater(FNAmplifications.getInstance(), () -> { // delay getting the actual hp of the damager
+                int playerDefaultHealth = (int) Objects.requireNonNull(player.getAttribute(Attribute.GENERIC_MAX_HEALTH)).getValue();
+                if (player.getHealth() <= playerDefaultHealth - 1 && !livingEntity.isDead()) {
+                    player.setHealth(player.getHealth() + 2);
+                    livingEntity.setHealth(livingEntity.getHealth() < 2 ? livingEntity.getHealth() + (livingEntity.getHealth() * (-1)) :
+                            livingEntity.getHealth() - 2);
+                    sendGemMessage(player, this.getItemName());
+                }
+            }, 3L);
+        }
     }
 
 }
