+++ src/main/java/ne/fnfal113/fnamplifications/utils/PlayerJoinLister.java
@@ -45,22 +45,15 @@ public class PlayerJoinLister implements Listener {
                 Utils.colorTranslator("&e&lFN &c&lAmpli&b&lfications &r&e" + FNAmplifications.getInstance().getDescription().getVersion()),
                 Utils.colorTranslator("&fChangelog"),
                 "",
-                Utils.colorTranslator("&d&l• &r&d&lUpgradeable Gems!"),
-                Utils.colorTranslator("  &bMost of the gems now has tiers from I to IV, where"),
-                Utils.colorTranslator("  &btrigger chances are increased on each tier, any gems"),
-                Utils.colorTranslator("  &byou have and socketed to items now default to tier I"),
-                Utils.colorTranslator("  &bwhich has the lower chance of being triggered and you"),
-                Utils.colorTranslator("  &bcan upgrade these gem through crafting and the use of"),
-                Utils.colorTranslator("  &bthe new multiblock FN Gem Upgrader"),
-                Utils.colorTranslator("  &bUpgradeable gem item is added in the Fn Gems category as"),
-                Utils.colorTranslator("  &ba guide which shows how to craft and upgrade existing"),
-                Utils.colorTranslator("  &bgems of same type and tier/level. Also check out the guide"),
-                Utils.colorTranslator("  &bfor the tier chances for each gems!"),
-                Utils.colorTranslator("&d&l• &r&d&lAdded &r&dFN Gem Upgrader Multiblock"),
-                Utils.colorTranslator("&d&l• &r&d&lFixed &r&dsome bugs"),
+                Utils.colorTranslator("&d&l• &r&d&lAdded Lifesteal Gem"),
+                Utils.colorTranslator("  &bsteal 1 heart from your enemy by chance"),
+                Utils.colorTranslator("  &bas long your health is not full and"),
+                Utils.colorTranslator("&d&l• &r&d&lRefactored &r&dconfiguration management (API)"),
+                Utils.colorTranslator("&d&l• &r&dCode refactoring and changes"),
                 Utils.colorTranslator(""),
-                Utils.colorTranslator("&eLifesteal gem to follow in the next build."),
-                Utils.colorTranslator("&eif there are bugs please report on github"),
+                Utils.colorTranslator("&eServer admins are advised to delete all of the"),
+                Utils.colorTranslator("&econfig files inside /plugins/FNAmplifications to"),
+                Utils.colorTranslator("&eregenerate new and clean config entries"),
                 Utils.colorTranslator("&ehttps://github.com/FN-FAL113"),
                 "||---------------------------------------------------||"
         );
